Update Restaurants and Payments repository

+++ Exebite.DataAccess/Repositories/PaymentRepository/IPaymentCommandRepository.cs
@@ -1,6 +1,6 @@
 ﻿namespace Exebite.DataAccess.Repositories
 {
-    public interface IPaymentCommandRepository : IDatabaseCommandRepository<int, PaymentInsertModel, PaymentUpdateModel>
+    public interface IPaymentCommandRepository : IDatabaseCommandRepository<long, PaymentInsertModel, PaymentUpdateModel>
     {
     }
 }
